Refactoring and cleanup of unused code.
