+++ src/app/message.rs
@@ -1,0 +1,7 @@
+use iced::widget::text_editor::Action;
+
+#[derive(Debug, Clone)]
+pub enum Message {
+    TextInputChanged(String),
+    TextEditorAction(Action),
+}
