--- src/view/mod.rs
@@ -1,4 +1,0 @@
-pub mod edview;
-pub mod mainview;
-pub mod svgview;
-pub mod tabview;
